requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0
flask-restx==1.3.0
Werkzeug==2.3.7